% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_data_Wald_test.R
\name{complete_data_Wald_test}
\alias{complete_data_Wald_test}
\title{The multivariate Wald test statistic for complete data.}
\usage{
complete_data_Wald_test(model, null_model)
}
\description{
The multivariate Wald test statistic for complete data.
}
