% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D0_naive_test.R
\name{D0_naive_test}
\alias{D0_naive_test}
\title{The most straightforward Wald statistic}
\usage{
D0_naive_test(model, null_model)
}
\description{
The most straightforward Wald statistic which does not need the
Equal Fraction of Missing Information" assumption.
}
