% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1_statistic.R
\name{D1_statistic}
\alias{D1_statistic}
\title{Compare two nested models using D1-statistic}
\usage{
D1_statistic(model, null_model)
}
\description{
The D1 statistic is the multivariate Wald test.
}
\references{
Li, K. H., T. E. Raghunathan, and D. B. Rubin. 1991.
Large-Sample Significance Levels from Multiply Imputed Data Using
Moment-Based Statistics and an F Reference Distribution.
\emph{Journal of the American Statistical Association}, 86(416): 1065-73.
}
