% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2_statistic.R
\name{D2_statistic}
\alias{D2_statistic}
\title{Compare two nested models using D2-statistic}
\usage{
D2_statistic(model, null_model)
}
\description{
The D2 statistic pools test statistics from the repeated analyses.
The method is less powerful than the D1 and D3 statistics.
}
\references{
Li, K. H., X. L. Meng, T. E. Raghunathan, and D. B. Rubin. 1991.
Significance Levels from Repeated p-Values with Multiply-Imputed Data.
\emph{Statistica Sinica} 1 (1): 65-92.
}
