% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_completed_data_moments.R
\name{get_completed_data_moments}
\alias{get_completed_data_moments}
\title{This function computes completed-data moments.}
\usage{
get_completed_data_moments(model, null_model, sandwich = FALSE)
}
\value{
A list containing the completed-data estimates and
their variance-covariance matrices.
}
\description{
(This is an internal function).
}
