% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_lrt_CM2019.R
\name{mi_lrt_CM2019}
\alias{mi_lrt_CM2019}
\title{Compare two nested models using D4 test}
\usage{
mi_lrt_CM2019(model, null_model)
}
\description{
This is the improved MI LRT by Chan and Meng (2019), of the original test by Meng and Rubin (1992).
}
