% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_lrt_MR1992.R
\name{mi_lrt_MR1992}
\alias{mi_lrt_MR1992}
\title{Compare two nested models using MI LRT}
\usage{
mi_lrt_MR1992(model, null_model)
}
\description{
The MI LRT (D3) test is a variant of likelihood ratio by Meng and Rubin (1992).
}
\references{
Meng, X. L., and D. B. Rubin. 1992.
Performing Likelihood Ratio Tests with Multiply-Imputed Data Sets.
\emph{Biometrika}, 79 (1): 103-11.
}
