% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mi_lrt_wilson_hilferty_(another_way_of_doing_same_thing).R
\name{mi_lrt_wilson_hilferty_another}
\alias{mi_lrt_wilson_hilferty_another}
\title{Compare two nested models}
\usage{
mi_lrt_wilson_hilferty_another(model, null_model, df = "Rubin1987")
}
\description{
This method is our proposed method using
the Wilson-Hilferty transformation of the LRT statistics
This codes uses the `mice::pool.scalar()` function. This does the same thing as `miLRT::mi_lrt_wilson_hilferty()`.
}
\references{
#' Meng, X. L., and D. B. Rubin. 1992.
Performing Likelihood Ratio Tests with Multiply-Imputed Data Sets.
\emph{Biometrika}, 79 (1): 103-11.

Wilson, E.B., and Hilferty, M.M. (1931).
The distribution of chi-square.
\emph{Proceedings of the National Academy of Sciences of the United States of America},
17, 684-688.
}
