% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_Wald_wilson_hilferty.R
\name{mi_Wald_wilson_hilferty}
\alias{mi_Wald_wilson_hilferty}
\title{Compare two nested models}
\usage{
mi_Wald_wilson_hilferty(model, null_model, df = "Rubin1987")
}
\description{
This method is our proposed method using
the Wilson-Hilferty transformation of the Wald statistics
}
\references{
Wilson, E.B., and Hilferty, M.M. (1931).
The distribution of chi-square.
\emph{Proceedings of the National Academy of Sciences of the United States of America},
17, 684-688.
}
