% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_std_norm.R
\name{transform_to_std_norm}
\alias{transform_to_std_norm}
\title{Returns the exact transformation of a chi-square random variable
to the standard normal random variable.}
\usage{
transform_to_std_norm(chi_square_rv, df)
}
\description{
Returns the exact transformation of a chi-square random variable
to the standard normal random variable.
}
