% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson_hilferty.R
\name{wilson_hilferty}
\alias{wilson_hilferty}
\title{Returns the Wilson-Hilferty transformation of a chi-square random variable}
\usage{
wilson_hilferty(chi_square_rv, df)
}
\description{
Returns the Wilson-Hilferty transformation of a chi-square random variable
}
\references{
Wilson, E.B., and Hilferty, M.M. (1931).
The distribution of chi-square.
\emph{Proceedings of the National Academy of Sciences of the United States of America},
17, 684-688.
}
